+++ resources/views/statistics.blade.php
@@ -1,4 +1,4 @@
-<div class="lingua">
+<x-lingua::dark-scope>
     <x-lingua::branding />
     <section class="flex flex-col gap-6">
 
@@ -40,7 +40,7 @@
         @endforeach
 
     </section>
-</div>
+</x-lingua::dark-scope>
 @assets
 @once
     <link rel="stylesheet" href="{{ linguaAssetUrl('css/lingua.min.css') }}">