+++ resources/views/translations.blade.php
@@ -1,4 +1,4 @@
-<div class="lingua">
+<x-lingua::dark-scope>
     <x-lingua::branding />
 
     <section class="flex flex-col gap-6">
@@ -225,7 +225,7 @@
             @endisland
         </div>
     </section>
-</div>
+</x-lingua::dark-scope>
 @assets
 @once
     <link rel="stylesheet" href="{{ linguaAssetUrl('css/lingua.min.css') }}">